requirements.txt
rich>=10.0.0
openai>=1.12.0
anthropic>=0.18.0
groq>=0.4.0
google-generativeai>=0.3.0
configparser>=5.0.0
crewai>=0.11.0
tavily-python>=0.1.9
langchain>=0.1.0
langchain-groq>=0.0.6
langchain-openai>=0.0.8
langchain-anthropic>=0.1.4
langchain-google-genai>=0.0.11
